refactor(config-ui): remove the connection submenu
