+++ config-ui/src/plugins/config.ts
@@ -17,6 +17,7 @@
  */
 
 import type { PluginConfigType } from './types';
+import { BasePipelineConfig } from './register/base';
 import { AEConfig } from './register/ae';
 import { AzureConfig } from './register/azure';
 import { BitBucketConfig } from './register/bitbucket';
@@ -38,9 +39,8 @@ import { SonarQubeConfig } from './register/sonarqube';
 import { StarRocksConfig } from './register/starrocks';
 import { TAPDConfig } from './register/tapd';
 import { WebhookConfig } from './register/webook';
-import { ZenTaoConfig } from './register/zentao';
 import { TeambitionConfig } from './register/teambition';
-import { BasePipelineConfig } from '@/plugins/register/base';
+import { ZenTaoConfig } from './register/zentao';
 
 export const PluginConfig: PluginConfigType[] = [
   AEConfig,
