+++ config-ui/src/plugins/register/github/config.tsx
@@ -16,10 +16,8 @@
  *
  */
 
-import React from 'react';
-
-import type { PluginConfigType } from '@/plugins';
-import { PluginType } from '@/plugins';
+import type { PluginConfigType } from '../../types';
+import { PluginType } from '../../types';
 
 import Icon from './assets/icon.svg';
 import { Token, Graphql } from './connection-fields';
