+++ config-ui/src/plugins/register/jenkins/config.ts
@@ -16,8 +16,8 @@
  *
  */
 
-import type { PluginConfigType } from '@/plugins';
-import { PluginType } from '@/plugins';
+import type { PluginConfigType } from '../../types';
+import { PluginType } from '../../types';
 
 import Icon from './assets/icon.svg';
